OCMCORE/requirements.txt
Django>=4.2
mssql-django>=1.3.0
gunicorn
waitress
drf-yasg
djangorestframework
pandas
python-memcached
django-pymemcache 